+++ project/src/settings/store-namespace.ts
@@ -1,0 +1,3 @@
+export enum StoreNamespace {
+  City = 'City',
+}
