+++ project/src/store/reducer.ts
@@ -1,12 +1,13 @@
 import {combineReducers} from '@reduxjs/toolkit';
-import {cityReducer} from 'store/city/reducer';
+import {cityReducer} from 'store/city/city-reducer';
 import {favoriteOffersReducer} from 'store/favorite-offers/reducer';
 import {offerReducer} from 'store/offer/reducer';
 import {offersReducer} from 'store/offers/reducer';
+import {StoreNamespace} from 'settings/store-namespace';
 import {userReducer} from 'store/user/reducer';
 
 export const reducer = combineReducers({
-  cityReducer,
+  [StoreNamespace.City]: cityReducer.reducer,
   favoriteOffersReducer,
   offerReducer,
   offersReducer,
